+++ monitor/core/taps/betabmp.py
@@ -140,6 +140,7 @@ def run_bgpstream_beta_bmp(prefixes_file=None):
                                     )
                             else:
                                 log.warning("Invalid format message: {}".format(msg))
+                            break
                 try:
                     elem = rec.get_next_elem()
                 except BaseException: