+++ monitor/core/taps/bgpstreamlive.py
@@ -154,6 +154,7 @@ def run_bgpstream(prefixes_file=None, projects=[], start=0, end=0):
                                     )
                             else:
                                 log.warning("Invalid format message: {}".format(msg))
+                            break
                 try:
                     elem = rec.get_next_elem()
                 except BaseException: